src/requirements.txt
telethon==1.39.0
PyYAML==6.0.2
requests==2.32.3
openai==1.65.2